Refactor "Run now" button into its own file
